+++ pkg/queryfrontend/roundtrip_test.go
@@ -13,17 +13,17 @@ import (
 	"testing"
 	"time"
 
-	"go.uber.org/atomic"
-
+	"github.com/efficientgo/core/testutil"
 	"github.com/go-kit/log"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/promql/parser"
 	"github.com/weaveworks/common/user"
+	"go.uber.org/atomic"
 
-	"github.com/efficientgo/core/testutil"
 	cortexcache "github.com/thanos-io/thanos/internal/cortex/chunk/cache"
 	"github.com/thanos-io/thanos/internal/cortex/cortexpb"
+	"github.com/thanos-io/thanos/internal/cortex/frontend/transport"
 	"github.com/thanos-io/thanos/internal/cortex/querier/queryrange"
 	cortexvalidation "github.com/thanos-io/thanos/internal/cortex/util/validation"
 	"github.com/thanos-io/thanos/pkg/store/labelpb"
@@ -186,6 +186,7 @@ func TestRoundTripRetryMiddleware(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			tpw, err := NewTripperware(
 				Config{
+					CortexHandlerConfig: &transport.HandlerConfig{},
 					QueryRangeConfig: QueryRangeConfig{
 						MaxRetries:             tc.maxRetries,
 						Limits:                 defaultLimits,
@@ -357,6 +358,7 @@ func TestRoundTripSplitIntervalMiddleware(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			tpw, err := NewTripperware(
 				Config{
+					CortexHandlerConfig: &transport.HandlerConfig{},
 					QueryRangeConfig: QueryRangeConfig{
 						Limits:                 defaultLimits,
 						SplitQueriesByInterval: tc.splitInterval,
@@ -461,6 +463,7 @@ func TestRoundTripQueryRangeCacheMiddleware(t *testing.T) {
 
 	tpw, err := NewTripperware(
 		Config{
+			CortexHandlerConfig: &transport.HandlerConfig{},
 			QueryRangeConfig: QueryRangeConfig{
 				Limits:                 defaultLimits,
 				ResultsCacheConfig:     cacheConf,
@@ -558,6 +561,7 @@ func TestRoundTripQueryCacheWithShardingMiddleware(t *testing.T) {
 				ResultsCacheConfig:     cacheConf,
 				SplitQueriesByInterval: day,
 			},
+			CortexHandlerConfig: &transport.HandlerConfig{},
 		}, nil, log.NewNopLogger(),
 	)
 	testutil.Ok(t, err)
@@ -682,6 +686,7 @@ func TestRoundTripLabelsCacheMiddleware(t *testing.T) {
 				ResultsCacheConfig:     cacheConf,
 				SplitQueriesByInterval: day,
 			},
+			CortexHandlerConfig: &transport.HandlerConfig{},
 		}, nil, log.NewNopLogger(),
 	)
 	testutil.Ok(t, err)
@@ -790,6 +795,7 @@ func TestRoundTripSeriesCacheMiddleware(t *testing.T) {
 
 	tpw, err := NewTripperware(
 		Config{
+			CortexHandlerConfig: &transport.HandlerConfig{},
 			LabelsConfig: LabelsConfig{
 				Limits:                 defaultLimits,
 				ResultsCacheConfig:     cacheConf,